requirements.txt
Flask==2.3.2
Jinja2==3.1.2
Werkzeug==2.3.6
python-dotenv==1.0.0
